authorization with gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -15,6 +15,11 @@ class AttendeeController extends Controller
 
     private array $relations = ['user'];
 
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -56,6 +61,9 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+
+        // argomenti multipli di authorize vanno in un array
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
